requirements.txt
aiohttp==3.8.3
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
oauth2client==4.1.3
oauthlib==3.2.0
paho-mqtt==1.6.1
paramiko==2.9.3
pyTelegramBotAPI==4.8.0
python-nmap==0.7.1
requests
save-thread-result==0.0.9
